# Rework several serialization benchmarks