--- a/src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Microbenchmarks/ResourceSet.cs
+++ b/src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Microbenchmarks/ResourceSet.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT license. See License.txt in the project root for license information.
+
+namespace Microsoft.AspNetCore.Razor.Microbenchmarks;
+
+public enum ResourceSet
+{
+    Legacy,
+    Telerik
+}