… src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Microbenchmarks/Resources/Telerik/Kendo.Mvc.Examples.taghelpers.json
… src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Microbenchmarks/Resources/project.razor.json
… src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Microbenchmarks/TagHelperBenchmarkBase.cs